Split voting methods into package modules
