+++ src/vse_sim/methods/__init__.py
@@ -1,0 +1,32 @@
+from .borda import Borda, RankedMethod, RatedMethod
+from .bullety_approval import BulletyApprovalWith
+from .irnr import IRNR
+from .irv import Irv
+from .irv_prime import IrvPrime
+from .mav import Mav, toVote
+from .mj import Mj
+from .plurality import Plurality
+from .ranked_pairs import Rp
+from .schulze import Schulze
+from .score import Score
+from .srv import Srv
+from .v321 import V321
+
+__all__ = [
+    "Borda",
+    "BulletyApprovalWith",
+    "IRNR",
+    "Irv",
+    "IrvPrime",
+    "Mav",
+    "Mj",
+    "Plurality",
+    "RankedMethod",
+    "RatedMethod",
+    "Rp",
+    "Schulze",
+    "Score",
+    "Srv",
+    "V321",
+    "toVote",
+]
